Fix and clean up DDCaloDigi to give the same results as the processor
and add a test to make sure this is happening.
The main changes that allow the algorithm to give the same result are
using the correct type for cellID (64 bits are needed in EDM4hep) and
storing `CalibrHCALOther` as a double instead of as a float.

Some of the cleanup actions were:
- Use const were possible
- Merge functions that were similar
- Remove unused variables
- Reduce the number of places where conversions happen (int to long, float to
double or viceversa, etc.)